requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
bcrypt==4.1.3
cryptography==42.0.7
PyJWT==2.8.0